check clinic staus
