+++ bootstrap/app.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Middleware\CheckClinicaStatus;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
@@ -7,54 +8,21 @@ use Illuminate\Auth\AuthenticationException;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
-//use Closure;
-
-/**
- * Middleware que verifica se o usuário (clinica) possui status "aprovado".
- */
-class CheckClinicaStatus
-{
-    public function handle(Request $request, Closure $next)
-    {
-        //Auth::guard('clinic')->user()->status;
-        $clinica = Auth::guard('clinic')->user(); // Obtém a clínica autenticada
-
-        $status = $clinica->status;
-
-        // Verifica o status e redireciona para a rota correspondente
-        switch ($status) {
-            case 'pendente':
-                // Redireciona para uma página informando que a clínica está pendente de aprovação
-                return redirect()->route('clinica.pendente');
-            case 'parcial':
-                // Redireciona para uma página que indica acesso negado.
-                return redirect()->route('clinica.negado');
-            case 'aprovado':
-                // Se estiver aprovado, permite o acesso normalmente
-                return $next($request);
-            default:
-                // Opcional: redireciona para uma página de erro ou logout
-                return redirect()->route('login');
-        }
-    }
-}
 
 return Application::configure(basePath: dirname(__DIR__))
     ->withRouting(
         web: __DIR__.'/../routes/web.php',
         api: __DIR__.'/../routes/api.php',
         commands: __DIR__.'/../routes/console.php',
-        health: '/up',
+        health: '/up'
     )
     ->withMiddleware(function (Middleware $middleware) {
         $middleware->redirectTo(
-            guests: '/login', // Rota para convidados (não autenticados)
+            guests: '/login',
             users: function ($request) {
-                // Se o usuário estiver autenticado com o guard 'clinic'
                 if (Auth::guard('clinic')->check()) {
                     return route('admin-clinica.dashboard.index');
                 }
-                // Rota padrão para outros guards
                 return route('perfil.edit');
             }
         );
@@ -66,19 +34,19 @@ return Application::configure(basePath: dirname(__DIR__))
 
         // Exclui a rota '/asaas/webhook' da verificação CSRF
         $middleware->validateCsrfTokens(except: [
-            'asaas/webhook', // Adicione outras URIs conforme necessário
+            'asaas/webhook',
         ]);
 
-        // Se desejar que ele seja aplicado globalmente, pode usá-lo com:
+        // Caso queira aplicar globalmente:
         // $middleware->append(CheckClinicaStatus::class);
     })
     ->withExceptions(function (Exceptions $exceptions) {
         $exceptions->renderable(function (AuthenticationException $e, $request) {
             $guards = $e->guards();
-            $guard = Arr::first($guards);
+            $guard  = Arr::first($guards);
             $loginRoute = match($guard) {
-                'clinic' => 'login2', // Rota para clínicas
-                default  => 'login'   // Rota padrão
+                'clinic' => 'login2',
+                default  => 'login'
             };
 
             return $request->expectsJson()
